refactor: 공통, 고양이 장난감 영역 패키지 분리
