+++ app/src/main/java/com/codesoom/assignment/cattoy/exception/CatToyInvalidFieldException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.exception;
+package com.codesoom.assignment.cattoy.exception;
 
 public class CatToyInvalidFieldException extends RuntimeException {
     public static final String DEFAULT_MESSAGE = "필드 정보를 공백 혹은 Null으로 수정할 수 없습니다.";
