+++ app/src/main/java/com/codesoom/assignment/cattoy/exception/CatToyInvalidPriceException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.exception;
+package com.codesoom.assignment.cattoy.exception;
 
 public class CatToyInvalidPriceException extends RuntimeException {
     public static final String DEFAULT_MESSAGE = "가격은 0원 이상이어야 합니다. 현재 입력 가격: [%d]";
