+++ app/src/main/java/com/codesoom/assignment/cattoy/exception/CatToyNotFoundException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.exception;
+package com.codesoom.assignment.cattoy.exception;
 
 public class CatToyNotFoundException extends RuntimeException {
     public static final String DEFAULT_MESSAGE = "식별자가 [%d]인 고양이 장난감을 찾을 수 없었습니다.";
