+++ app/src/main/java/com/codesoom/assignment/cattoy/exception/NotSupportedTypeException.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.exception;
+package com.codesoom.assignment.cattoy.exception;
 
 public class NotSupportedTypeException extends RuntimeException {
     public static final String DEFAULT_MESSAGE = "유효하지 않은 타입입니다. 지원 타입: [%s] , 입력 타입: [%s]";
