+++ app/src/test/java/com/codesoom/assignment/ProvideInvalidCatToyArguments.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment;
 
-import com.codesoom.assignment.domain.CatToy;
+import com.codesoom.assignment.cattoy.domain.CatToy;
 import org.junit.jupiter.api.extension.ExtensionContext;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.ArgumentsProvider;
